Enabling Transaction in MongoDB
For this we have to use EnableTransactionManagement and then provide an implementation for  PlatformTransactionManager. In our case as we're using mongodb we'll provide the inbuilt config using MongoTransactionManager

Note: in mongodb replication is mandatory for transaction to happen, we don't need to do anything about this since the database is already managed by MongoDB Atlas.