using yml jlt
Application properties support only strings as values, whereas Application YAML supports multiple data types, including numbers, booleans, arrays, and objects. This makes YAML more flexible and suitable for applications with complex configuration requirements. If both the files are present and the properties are same but the values differ, then the priority is given to application.properties